Add unsafeSpliceCoerce (#6)
This is one component used in `liftTypedFromUntypedSplice`. It is
useful to have `unsafeSpliceCoerce` as its own function because while
`liftTypedFromUntypedSplice` cannot be levity polymorphic (due to GHC
restrictions), `unsafeSpliceCoerce` can be levity polymorphic. This
makes `unsafeSpliceCoerce` useful for manully implementing
`liftTyped` for unlifted `Lift` instances in a backwards compatible
way. (See the `unsafeSpliceCoerce` Haddocks for an example of how to
do this.)